test(sdk): consolidate ELB discovery integration coverage
