+++ docs/reference/rule-ids.md
@@ -118,7 +118,7 @@ meaningful optimization opportunities, and `low` covers hygiene and smaller accu
 | `CLDBRN-AWS-ELB-2`                 | medium   | Flag Classic Load Balancers that have zero attached instances.                                                                                                                                                                              | elb                 | discovery      |
 | `CLDBRN-AWS-ELB-3`                 | medium   | Flags load balancers with no attached target groups or no registered targets across attached target groups.                                                                                                                                 | elb                 | discovery      |
 | `CLDBRN-AWS-ELB-4`                 | medium   | Flags load balancers with no attached target groups or no registered targets across attached target groups.                                                                                                                                 | elb                 | discovery      |
-| `CLDBRN-AWS-ELB-5`                 | medium   | Requires a complete 14-day `RequestCount` history, treats fewer than `10` requests per day as idle, and skips load balancers already covered by the stricter empty-target cleanup rules.                                                    | elb                 | discovery      |
+| `CLDBRN-AWS-ELB-5`                 | medium   | Flags ALBs and HTTP/HTTPS-only Classic load balancers below `10` requests/day with complete 14-day evidence. Unsupported or incomplete activity remains unknown; empty-target cleanup takes precedence.                                                    | elb                 | discovery      |
 | `CLDBRN-AWS-EMR-1`                 | medium   | Reuses the built-in EC2 family policy. EMR clusters are flagged when any discovered cluster instance type falls into the current non-preferred, previous-generation family set.                                                             | emr                 | discovery, iac |
 | `CLDBRN-AWS-EMR-2`                 | high     | Flags only active clusters whose `IsIdle` metric stays true for six consecutive 5-minute periods, which is a 30-minute idle window.                                                                                                         | emr                 | discovery      |
 | `CLDBRN-AWS-RDS-1`                 | medium   | Flag RDS DB instances that do not use curated preferred instance classes.                                                                                                                                                                   | rds                 | iac, discovery |
